% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchoredDistr.R
\docType{package}
\name{anchoredDistr}
\alias{anchoredDistr}
\alias{anchoredDistr-package}
\title{anchoredDistr: A package for post-processing of the Method of Anchored Distributions.}
\description{
The anchoredDistr package provides functions for post-processing applications of
the Method of Anchored Distributions (MAD) such as viewing data, calculating
likelihood and posterior distributions,
testing convergence of likelihood values, and applying dimension reduction.
}
