% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.MADproject}
\alias{print.MADproject}
\title{Print the data contained in MADproject object slots.}
\usage{
\method{print}{MADproject}(x, ...)
}
\arguments{
\item{x}{The MADproject object}

\item{...}{Not supported}
}
\description{
\code{print.MADproject} plots information on the contents
in the slots of the provided MADproject.
}
\examples{
data(pumping)
print(pumping)

}
