% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pumping}
\alias{pumping}
\title{anchoredDistr example data}
\format{A vector of observations of indirect data, a
data.frame of prior distribution samples, and a
data.frame of simualted realizations of indirect data
based on the samples.

\describe{
  \item{obs}{a time series of drawdown, in meters}
  \item{priors}{a data.frame of prior samples and their probability density}
  \item{realizations}{the ensemble of simulated drawdown time series for the samples}
}}
\usage{
pumping

pumping
}
\description{
A dataset containing data to create a MADproject object
called 'pumping' without a MAD# database

A dataset containing a MADproject object called 'pumping'
}
\keyword{datasets}

