% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tutorial}
\alias{tutorial}
\title{anchoredDistr example data}
\format{A MADproject object with an observed head at three wells, 145
prior samples, realizations for each sample, and true values for
validation where the parameters to be inferred are four anchors of the
local decimal log transmissivity of a 1D aquifer
\describe{
  \item{@observations}{three head measurements, in meters}
  \item{@realizations}{the ensemble of simulated head measurements}
  \item{@numAnchors}{the value 4}
  \item{@priors}{the samples for each anchor's prior distribution
  based on the kriging mean and variance at the anchor locations}
}}
\source{
\url{http://www.mad.codeplex.com/releases/}
}
\usage{
data(tutorial)
}
\description{
A dataset containing a MADproject object called 'tutorial'
}
\keyword{datasets}
