% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pumping}
\alias{pumping}
\title{anchoredDistr example data}
\format{A MADproject object with an observed time series, three
prior samples, realizations for each sample, and true values for
validation where the parameter to be inferred is the global mean
of the natural log conductivity of a 2D aquifer.

\describe{
  \item{@observations}{a time series of drawdown, in meters}
  \item{@numSamples}{the value 3, [unitless]}
  \item{@realizations}{the ensemble of simulated drawdown time series}
}}
\usage{
pumping
}
\description{
A dataset containing a MADproject object called 'pumping'
}
\keyword{datasets}

